requirements.txt
groq>=0.11.0
google-generativeai>=0.8.0

sentence-transformers>=3.0.0
chromadb>=0.5.0

langchain-text-splitters>=0.3.0
langgraph>=0.2.0
rank-bm25>=0.2.2

fastapi>=0.115.0
uvicorn>=0.30.0
slowapi>=0.1.9

presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

structlog>=24.0.0

python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.66.0

matplotlib>=3.8.0
